requirements.txt
Flask
PyJWT
Flask-MySQLdb
passlib